replication: force gc to clean logs on no disk space err
Garbage collector do not delete xlog unless replica do not notify
master with newer vclock. This can lead to running out of disk
space error and this is not right behaviour since it will stop the
master.

Add error injection and a test.

WARNING: This is design proposal, not ready for commit!!!

Proposal for #3397